mv JSI::SchemaClasses to jsi/schema_classes.rb
